+++ packages/dive/lib/src/dive_media_source.dart
@@ -1,10 +1,13 @@
 import 'dart:async';
 import 'package:dive_obslib/dive_obslib.dart';
+import 'package:equatable/equatable.dart';
 import 'package:riverpod/riverpod.dart';
 
 import 'dive_audio_meter_source.dart';
 import 'dive_core.dart';
 import 'dive_input_type.dart';
+import 'dive_settings.dart';
+import 'dive_settings_data.dart';
 import 'dive_sources.dart';
 
 /// The media playback states.
@@ -20,7 +23,7 @@ enum DiveMediaState {
 }
 
 /// The state model for a media source.
-class DiveMediaSourceState {
+class DiveMediaSourceState extends Equatable {
   /// The current playback time.
   final int currentTime;
 
@@ -30,12 +33,16 @@ class DiveMediaSourceState {
   /// The current media playback state.
   final DiveMediaState mediaState;
 
-  DiveMediaSourceState({this.currentTime = 0, this.duration = 0, this.mediaState = DiveMediaState.NONE});
+  const DiveMediaSourceState(
+      {this.currentTime = 0, this.duration = 0, this.mediaState = DiveMediaState.NONE});
 
+  /// Indicates to output this instance's [props] in [toString].
   @override
-  String toString() {
-    return "DiveMediaSourceState: time=$currentTime, duration=$duration, mediaState=$mediaState";
-  }
+  bool get stringify => true;
+
+  /// Returns the list of properties used to facilitate [operator ==].
+  @override
+  List<Object> get props => [currentTime, duration, mediaState];
 }
 
 class _DiveMediaSourceStateNotifier extends StateNotifier<DiveMediaSourceState> {
@@ -44,35 +51,89 @@ class _DiveMediaSourceStateNotifier extends StateNotifier<DiveMediaSourceState> 
   _DiveMediaSourceStateNotifier(DiveMediaSourceState stateModel)
       : super(stateModel ?? DiveMediaSourceState());
 
-  void updateMediaState(DiveMediaSourceState stateModel) {
-    state = stateModel;
+  void updateMediaState(DiveMediaSourceState stateModel) => state = stateModel;
+}
+
+class DiveMediaSourceSettings extends DiveSettings {
+  String get localFile => get('local_file');
+  bool get isLocalFile => get('is_local_file');
+  bool get looping => get('looping');
+
+  bool get clearOnMediaEnd => get('clear_on_media_end');
+  bool get restartOnActivate => get('restart_on_activate');
+  bool get linearAlpha => get('linear_alpha');
+
+  int get reconnectDelaySec => get('reconnect_delay_sec');
+  int get bufferingMb => get('buffering_mb');
+  int get speedPercent => get('speed_percent');
+
+  DiveMediaSourceSettings({
+    String localFile,
+    bool isLocalFile = true,
+    bool looping = false,
+    bool clearOnMediaEnd = true,
+    bool restartOnActivate = true,
+    bool linearAlpha = false,
+    int reconnectDelaySec = 10,
+    int bufferingMb = 2,
+    int speedPercent = 100,
+  }) {
+    set('local_file', localFile);
+    set('is_local_file', isLocalFile);
+    set('looping', looping);
+    set('clear_on_media_end', clearOnMediaEnd);
+    set('restart_on_activate', restartOnActivate);
+    set('linear_alpha', linearAlpha);
+    set('reconnect_delay_sec', reconnectDelaySec);
+    set('buffering_mb', bufferingMb);
+    set('speed_percent', speedPercent);
   }
 }
 
+/// Represents a ffmpeg_source media source.
 class DiveMediaSource extends DiveTextureSource {
   final stateProvider =
       StateNotifierProvider<_DiveMediaSourceStateNotifier>((ref) => _DiveMediaSourceStateNotifier(null));
 
-  Timer _playbackTimer;
-
-  final String localFile;
+  // final String localFile;
+  final DiveMediaSourceSettings settings;
 
   DiveAudioMeterSource volumeMeter;
 
-  DiveMediaSource({String name, this.localFile}) : super(inputType: DiveInputType.mediaSource, name: name) {
-    _playbackTimer = Timer.periodic(Duration(milliseconds: 1000), timerCallback);
-  }
+  /// The state syncing interval (msec).
+  final stateSyncInterval;
 
-  static Future<DiveMediaSource> create(String localFile) async {
-    final source = DiveMediaSource(name: 'my media', localFile: localFile);
-    await source.setupTexture(source.trackingUUID);
-    source.pointer = obslib.createMediaSource(source.trackingUUID, localFile);
+  Timer _syncStateTimer;
+
+  DiveMediaSource(String name, {this.settings, this.stateSyncInterval = 500})
+      : super(inputType: DiveInputType.mediaSource, name: name) {
+    if (stateSyncInterval > 0) {
+      _syncStateTimer = Timer.periodic(Duration(milliseconds: stateSyncInterval), _timerCallback);
+    }
+  }
+  static Future<DiveMediaSource> create({
+    String name = 'my media',
+    DiveMediaSourceSettings settings,
+    bool requiresVideoMonitor = true,
+  }) async {
+    final source = DiveMediaSource(name, settings: settings);
+    if (requiresVideoMonitor) await source.setupTexture(source.trackingUUID);
+
+    final data = settings.toData();
+    source.pointer = obslib.createMediaSource(
+      sourceUuid: source.trackingUUID,
+      name: name,
+      settings: data,
+    );
+
+    data.dispose();
     await source.syncState();
 
     // TODO: try adding frame callback
     // await obslib.addSourceFrameCallback(source.trackingUUID, source.pointer.address);
 
     // Turn on the default audio monitoring device so we can hear the audio.
+    // TODO: is this really needed here?
     if (source.pointer != null) {
       obslib.sourceSetMonitoringType(source.pointer);
     }
@@ -83,9 +144,9 @@ class DiveMediaSource extends DiveTextureSource {
   /// Release the resources associated with this source.
   @override
   bool dispose() {
-    if (_playbackTimer != null) {
-      _playbackTimer.cancel();
-      _playbackTimer = null;
+    if (_syncStateTimer != null) {
+      _syncStateTimer.cancel();
+      _syncStateTimer = null;
     }
 
     // obslib.removeSourceFrameCallback(trackingUUID, pointer.address);
@@ -101,10 +162,8 @@ class DiveMediaSource extends DiveTextureSource {
     return true;
   }
 
-  void timerCallback(Timer timer) {
-    if (DiveCore.notifierFor(stateProvider).stateModel.mediaState == DiveMediaState.PLAYING) {
-      _syncState();
-    }
+  void _timerCallback(Timer timer) {
+    _syncState();
   }
 
   /// Sync the media state from the media source to the state provider,
@@ -121,8 +180,11 @@ class DiveMediaSource extends DiveTextureSource {
 
   /// Sync the media state from the media source to the state provider.
   Future<void> _syncState() async {
-    final state = await getState();
-    DiveCore.notifierFor(stateProvider).updateMediaState(state);
+    final currentStateModel = DiveCore.notifierFor(stateProvider).stateModel;
+    final newStateModel = await getState();
+    if (currentStateModel != newStateModel) {
+      DiveCore.notifierFor(stateProvider).updateMediaState(newStateModel);
+    }
     return;
   }
 
@@ -133,13 +195,6 @@ class DiveMediaSource extends DiveTextureSource {
     final ms = obslib.mediaSourceGetTime(pointer);
     return DiveMediaSourceState(
         currentTime: ms, duration: duration, mediaState: DiveMediaState.values[mediaState]);
-    // return DivePlugin.mediaGetState(trackingUUID).then((mediaState) =>
-    //     DivePlugin.mediaGetDuration(trackingUUID).then((duration) =>
-    //         DivePlugin.mediaGetTime(trackingUUID).then((ms) =>
-    //             DiveMediaSourceState(
-    //                 currentTime: ms,
-    //                 duration: duration,
-    //                 mediaState: DiveMediaState.values[mediaState]))));
   }
 
   Future<bool> play() async {
@@ -153,48 +208,30 @@ class DiveMediaSource extends DiveTextureSource {
     obslib.mediaSourcePlayPause(pointer, true);
     await syncState();
     return true;
-    // final rv = await DivePlugin.mediaPlayPause(trackingUUID, true);
-    // if (rv) {
-    //   await syncState();
-    // }
-    // return rv;
   }
 
   Future<bool> restart() async {
     obslib.mediaSourceRestart(pointer);
     await syncState();
     return true;
-    // final rv = await DivePlugin.mediaRestart(trackingUUID);
-    // if (rv) {
-    //   await syncState();
-    // }
-    // return rv;
   }
 
   Future<bool> stop() async {
     obslib.mediaSourceStop(pointer);
     await syncState();
     return true;
-    // final rv = await DivePlugin.mediaStop(trackingUUID);
-    // if (rv) {
-    //   await syncState();
-    // }
-    // return rv;
   }
 
   Future<int> getDuration() async {
     return obslib.mediaSourceGetDuration(pointer);
-    // return DivePlugin.mediaGetDuration(trackingUUID);
   }
 
   Future<int> getTime() async {
     return obslib.mediaSourceGetTime(pointer);
-    // return DivePlugin.mediaGetTime(trackingUUID);
   }
 
   Future<bool> setTime(int ms) async {
     obslib.mediaSourceSetTime(pointer, ms);
     return true;
-    // return DivePlugin.mediaSetTime(trackingUUID, ms);
   }
 }